+++ aom_dsp/bitwriter_buffer.c
@@ -90,10 +90,10 @@ void aom_wb_write_uvlc(struct aom_write_bit_buffer *wb, uint32_t v) {
 }
 
 void aom_wb_write_primitive_quniform(struct aom_write_bit_buffer *wb,
-                                     uint16_t n, uint16_t v) {
+                                     uint32_t n, uint32_t v) {
   if (n <= 1) return;
-  const int l = get_msb(n) + 1;
-  const int m = (1 << l) - n;
+  const uint32_t l = get_msb(n) + 1;
+  const uint32_t m = (1 << l) - n;
   if (v < m) {
     aom_wb_write_literal(wb, v, l - 1);
   } else {
@@ -103,17 +103,17 @@ void aom_wb_write_primitive_quniform(struct aom_write_bit_buffer *wb,
 }
 
 static void wb_write_primitive_subexpfin(struct aom_write_bit_buffer *wb,
-                                         uint16_t n, uint16_t k, uint16_t v) {
-  int i = 0;
-  int mk = 0;
+                                         uint32_t n, uint32_t k, uint32_t v) {
+  uint32_t i = 0;
+  uint32_t mk = 0;
   while (1) {
-    int b = (i ? k + i - 1 : k);
-    int a = (1 << b);
+    uint32_t b = (i ? k + i - 1 : k);
+    uint32_t a = (1 << b);
     if (n <= mk + 3 * a) {
       aom_wb_write_primitive_quniform(wb, n - mk, v - mk);
       break;
     } else {
-      int t = (v >= mk + a);
+      uint32_t t = (v >= mk + a);
       aom_wb_write_bit(wb, t);
       if (t) {
         i = i + 1;
@@ -127,16 +127,16 @@ static void wb_write_primitive_subexpfin(struct aom_write_bit_buffer *wb,
 }
 
 static void wb_write_primitive_refsubexpfin(struct aom_write_bit_buffer *wb,
-                                            uint16_t n, uint16_t k,
-                                            uint16_t ref, uint16_t v) {
+                                            uint32_t n, uint32_t k,
+                                            uint32_t ref, uint32_t v) {
   wb_write_primitive_subexpfin(wb, n, k, recenter_finite_nonneg(n, ref, v));
 }
 
 void aom_wb_write_signed_primitive_refsubexpfin(struct aom_write_bit_buffer *wb,
-                                                uint16_t n, uint16_t k,
-                                                int16_t ref, int16_t v) {
+                                                uint32_t n, uint32_t k,
+                                                int32_t ref, int32_t v) {
   ref += n - 1;
   v += n - 1;
-  const uint16_t scaled_n = (n << 1) - 1;
+  const uint32_t scaled_n = (n << 1) - 1;
   wb_write_primitive_refsubexpfin(wb, scaled_n, k, ref, v);
 }
